Rename e2e dirs for clarity
- e2e/ -> e2e-mock-api/
- e2e-demo/ -> e2e-demo-mode/
- Update all scripts, configs, and CI references
