+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
       - run: npm run lint
       - run: npm run build
 
-  test-e2e:
+  test-e2e-mock-api:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
@@ -35,9 +35,9 @@ jobs:
       - run: npm ci
       - run: npm run generate
       - run: npx playwright install --with-deps chromium
-      - run: npm run test:e2e
+      - run: npm run test:e2e:mock-api
 
-  test-e2e-demo:
+  test-e2e-demo-mode:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
@@ -50,7 +50,7 @@ jobs:
       - run: npm ci
       - run: npm run generate
       - run: npx playwright install --with-deps chromium
-      - run: npm run test:e2e:demo
+      - run: npm run test:e2e:demo-mode
 
   test-integration:
     runs-on: ubuntu-latest
